src/koopman_rbc/utils/boilerplate/clean_memory.py
import gc
from typing import Any

from hydra.core.utils import JobReturn
from hydra.experimental.callback import Callback
from omegaconf import DictConfig


class CleanMemoryCallback(Callback):
    def on_job_end(
        self, config: DictConfig, job_return: JobReturn, **kwargs: Any
    ) -> None:
        gc.collect()
