pyproject.toml
[project]
name = "koopman-rbc"
version = "1.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "derivative>=0.6.3",
    "ffmpeg>=1.4",
    "h5py>=3.13.0",
    "hydra-colorlog>=1.2.0",
    "hydra-core>=1.3.2",
    "lightning>=2.5.1",
    "matplotlib>=3.10.1",
    "optht>=0.2.0",
    "pandas>=2.2.3",
    "pydmd>=2025.4.1",
    "pykoopman>=1.0.5",
    "rich>=14.0.0",
    "seaborn>=0.13.2",
    "torch>=2.6.0",
    "torchmetrics>=1.7.1",
    "torchvision>=0.21.0",
    "wandb>=0.19.9",
]

[dependency-groups]
dev = [
    "ruff>=0.11.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
